back-end2/requirements.txt

fastapi==0.100.0
uvicorn[standard]==0.22.0
asyncpg==0.27.0
pydantic==2.5.1
python-dotenv==1.0.0

